MaxLock - No Xposed Implementation
This is a mostly stable, but not fully finished implementation of MaxLock which doesn't require root or Xposed and therefore also works on Android Nougat and O (tested on Preview!). It uses an accessibility service to monitor window state changes, and also automatically relocks if killed (and restarted). Further modifications, maybe even using root, will hopefully improve the reliability and security of this implementation.